+++ src/features/MainScreen/index.tsx
@@ -2,13 +2,13 @@ import React, { FC, useEffect } from 'react';
 import { Panel, PanelResizeHandle } from 'react-resizable-panels';
 import { Box, HStack, VStack } from '@chakra-ui/react';
 import { SyncedPanelGroup } from '@components/SyncedPanelGroup';
+import { useWorkspaceShortcutsHandlers } from '@features/App/Workspace/useWorkspaceShortcutsHandlers';
 import { useTagsRegistry } from '@features/App/Workspace/WorkspaceProvider';
 import { NotesListPanel } from '@features/MainScreen/NotesListPanel';
 import { WorkspacesPanel } from '@features/MainScreen/WorkspacesPanel';
 import { NotesContainer } from '@features/NotesContainer';
 import { useNoteCommandHandlers } from '@hooks/notes/useNoteCommandHandlers';
 import { useUpdateNotes } from '@hooks/notes/useUpdateNotes';
-import { useWorkspaceShortcutsHandlers } from '@hooks/shortcuts/useWorkspaceShortcutsHandlers';
 
 import { NotificationsPopup } from '../NotificationsPopup';
 import { ActivityBar } from './ActivityBar';
